recog in rasmm

+++ AFQ/tasks/tractography.py
@@ -88,7 +88,7 @@ def streamlines(data_imap, seed, tissue_imap, fodf, citations, tracking_params):
 
     if is_trx:
         start_time = time()
-        dtype_dict = {"positions": np.float16, "offsets": np.uint32}
+        dtype_dict = {"positions": np.float32, "offsets": np.uint32}
         if num_chunks and num_chunks > 1:
             if not has_ray:
                 raise ImportError(
